editregions/requirements.txt
six>=1.3.0
django-appconf>=0.6
